Use a map to back meta's cache.
This is a first step to making the baking store of `Meta`'s underlying
data structures swappable. The first implementation is
using a simple array of key value pairs, but since we are using the
`set`, `get`, `has`, `delete` interface we can change the backing data
store at will (or even "upgrade" at runtime).

The underlying implementation of the "lodash-stack" is mostly copied
from lodash itself (with inline attribution), but in the longer term we
should simply use this from lodash itself (and embed these modules from
lodash-es in our build process).